APP-1316: Parcels: Update SDK in accordance to Itella endpoints

+++ src/main/kotlin/com/paysera/lib/parcel/entities/PSCellSize.kt
@@ -2,7 +2,7 @@ package com.paysera.lib.parcel.entities
 
 data class PSCellSize(
     val name: String,
-    val width: Int,
-    val height: Int,
-    val depth: Int
+    val width: Double,
+    val height: Double,
+    val depth: Double
 )
